+++ src/language-handlebars/index.js
@@ -1,9 +1,13 @@
 import createParsers from "../utils/create-parsers.js";
-import { default as parsersConfig } from "./parsers.js";
 import printer from "./printer-glimmer.js";
 
 export const printers = {
   glimmer: printer,
 };
-export const parsers = createParsers(parsersConfig);
+export const parsers = createParsers([
+  {
+    importParsers: () => import("./parser-glimmer.js"),
+    parserNames: ["glimmer"],
+  },
+]);
 export { default as languages } from "./languages.evaluate.js";