+++ src/language-html/embed.js
@@ -28,17 +28,18 @@ import {
   inferElementParser,
   htmlTrimPreserveIndentation,
   dedentString,
-  unescapeQuoteEntities,
   isVueSlotAttribute,
   isVueSfcBindingsAttribute,
   getTextValueParts,
+  getUnescapedAttributeValue,
 } from "./utils/index.js";
+import isVueSfcWithTypescriptScript from "./utils/is-vue-sfc-with-typescript-script.js";
 import getNodeContent from "./get-node-content.js";
 
-async function printEmbeddedAttributeValue(node, htmlTextToDoc, options) {
+async function printEmbeddedAttributeValue(path, htmlTextToDoc, options) {
+  const { node } = path;
   const isKeyMatched = (patterns) =>
     new RegExp(patterns.join("|")).test(node.fullName);
-  const getValue = () => unescapeQuoteEntities(node.value);
 
   let shouldHug = false;
 
@@ -57,7 +58,8 @@ async function printEmbeddedAttributeValue(node, htmlTextToDoc, options) {
       rootNode &&
       (rootNode.type === "ObjectExpression" ||
         rootNode.type === "ArrayExpression" ||
-        (options.parser === "__vue_expression" &&
+        ((options.parser === "__vue_expression" ||
+          options.parser === "__vue_ts_expression") &&
           (rootNode.type === "TemplateLiteral" ||
             rootNode.type === "StringLiteral")))
     ) {
@@ -76,40 +78,43 @@ async function printEmbeddedAttributeValue(node, htmlTextToDoc, options) {
       __embeddedInHtml: true,
       ...opts,
     });
+  const value = getUnescapedAttributeValue(node);
 
   if (
     node.fullName === "srcset" &&
     (node.parent.fullName === "img" || node.parent.fullName === "source")
   ) {
-    return printExpand(printImgSrcset(getValue()));
+    return printExpand(printImgSrcset(value));
   }
 
-  if (node.fullName === "class" && !options.parentParser) {
-    const value = getValue();
-    if (!value.includes("{{")) {
-      return printClassNames(value);
-    }
+  if (
+    node.fullName === "class" &&
+    !options.parentParser &&
+    !value.includes("{{")
+  ) {
+    return printClassNames(value);
   }
 
-  if (node.fullName === "style" && !options.parentParser) {
-    const value = getValue();
-    if (!value.includes("{{")) {
-      return printExpand(
-        await attributeTextToDoc(value, {
-          parser: "css",
-          __isHTMLStyleAttribute: true,
-        })
-      );
-    }
+  if (
+    node.fullName === "style" &&
+    !options.parentParser &&
+    !value.includes("{{")
+  ) {
+    return printExpand(
+      await attributeTextToDoc(value, {
+        parser: "css",
+        __isHTMLStyleAttribute: true,
+      })
+    );
   }
 
   if (options.parser === "vue") {
     if (node.fullName === "v-for") {
-      return printVueFor(getValue(), attributeTextToDoc);
+      return printVueFor(path, attributeTextToDoc, options);
     }
 
     if (isVueSlotAttribute(node) || isVueSfcBindingsAttribute(node, options)) {
-      return printVueBindings(getValue(), attributeTextToDoc);
+      return printVueBindings(path, attributeTextToDoc, options);
     }
 
     /**
@@ -130,10 +135,11 @@ async function printEmbeddedAttributeValue(node, htmlTextToDoc, options) {
     const jsExpressionBindingPatterns = ["^v-"];
 
     if (isKeyMatched(vueEventBindingPatterns)) {
-      const value = getValue();
       const parser = isVueEventBindingExpression(value)
-        ? "__js_expression"
-        : options.__should_parse_vue_template_with_ts
+        ? isVueSfcWithTypescriptScript(path, options)
+          ? "__ts_expression"
+          : "__js_expression"
+        : isVueSfcWithTypescriptScript(path, options)
         ? "__vue_ts_event_binding"
         : "__vue_event_binding";
       return printMaybeHug(await attributeTextToDoc(value, { parser }));
@@ -141,13 +147,21 @@ async function printEmbeddedAttributeValue(node, htmlTextToDoc, options) {
 
     if (isKeyMatched(vueExpressionBindingPatterns)) {
       return printMaybeHug(
-        await attributeTextToDoc(getValue(), { parser: "__vue_expression" })
+        await attributeTextToDoc(value, {
+          parser: isVueSfcWithTypescriptScript(path, options)
+            ? "__vue_ts_expression"
+            : "__vue_expression",
+        })
       );
     }
 
     if (isKeyMatched(jsExpressionBindingPatterns)) {
       return printMaybeHug(
-        await attributeTextToDoc(getValue(), { parser: "__js_expression" })
+        await attributeTextToDoc(value, {
+          parser: isVueSfcWithTypescriptScript(path, options)
+            ? "__ts_expression"
+            : "__js_expression",
+        })
       );
     }
   }
@@ -185,33 +199,29 @@ async function printEmbeddedAttributeValue(node, htmlTextToDoc, options) {
     const ngI18nPatterns = ["^i18n(-.+)?$"];
 
     if (isKeyMatched(ngStatementBindingPatterns)) {
-      return printMaybeHug(
-        await ngTextToDoc(getValue(), { parser: "__ng_action" })
-      );
+      return printMaybeHug(await ngTextToDoc(value, { parser: "__ng_action" }));
     }
 
     if (isKeyMatched(ngExpressionBindingPatterns)) {
       return printMaybeHug(
-        await ngTextToDoc(getValue(), { parser: "__ng_binding" })
+        await ngTextToDoc(value, { parser: "__ng_binding" })
       );
     }
 
     if (isKeyMatched(ngI18nPatterns)) {
-      const value = getValue().trim();
       return printExpand(
-        fill(getTextValueParts(node, value)),
+        fill(getTextValueParts(node, value.trim())),
         !value.includes("@@")
       );
     }
 
     if (isKeyMatched(ngDirectiveBindingPatterns)) {
       return printMaybeHug(
-        await ngTextToDoc(getValue(), { parser: "__ng_directive" })
+        await ngTextToDoc(value, { parser: "__ng_directive" })
       );
     }
 
     const interpolationRegex = /{{(.+?)}}/s;
-    const value = getValue();
     if (interpolationRegex.test(value)) {
       const parts = [];
       for (const [index, part] of value.split(interpolationRegex).entries()) {
@@ -330,10 +340,12 @@ function embed(path, options) {
             textToDocOptions.parser = "__ng_interpolation";
             textToDocOptions.trailingComma = "none";
           } else if (options.parser === "vue") {
-            textToDocOptions.parser =
-              options.__should_parse_vue_template_with_ts
-                ? "__vue_ts_expression"
-                : "__vue_expression";
+            textToDocOptions.parser = isVueSfcWithTypescriptScript(
+              path,
+              options
+            )
+              ? "__vue_ts_expression"
+              : "__vue_expression";
           } else {
             textToDocOptions.parser = "__js_expression";
           }
@@ -383,7 +395,7 @@ function embed(path, options) {
 
       return async (textToDoc) => {
         const embeddedAttributeValueDoc = await printEmbeddedAttributeValue(
-          node,
+          path,
           (code, opts) =>
             // strictly prefer single quote to avoid unnecessary html entity escape
             textToDoc(code, {