+++ src/language-js/parse/espree.js
@@ -55,4 +55,4 @@ function parse(text, options = {}) {
   return postprocess(ast, { text });
 }
 
-export default createParser(parse);
+export const espree = createParser(parse);