+++ src/language-js/parse/postprocess/typescript.js
@@ -45,6 +45,18 @@ function throwErrorForInvalidAbstractProperty(tsNode, esTreeNode) {
   );
 }
 
+function nodeCanBeDecorated(node) {
+  return [true, false].some((useLegacyDecorators) =>
+    // @ts-expect-error -- internal?
+    ts.nodeCanBeDecorated(
+      useLegacyDecorators,
+      node,
+      node.parent,
+      node.parent.parent
+    )
+  );
+}
+
 // Based on `checkGrammarModifiers` function in `typescript`
 function throwErrorForInvalidModifier(node) {
   const { modifiers } = node;
@@ -55,50 +67,19 @@ function throwErrorForInvalidModifier(node) {
   const { SyntaxKind } = ts;
 
   for (const modifier of modifiers) {
-    if (ts.isDecorator(modifier)) {
-      const legacyDecorators = true;
+    if (ts.isDecorator(modifier) && !nodeCanBeDecorated(node)) {
       if (
+        node.kind === SyntaxKind.MethodDeclaration &&
         // @ts-expect-error -- internal?
-        !ts.nodeCanBeDecorated(
-          legacyDecorators,
-          node,
-          node.parent,
-          node.parent.parent
-        )
+        !ts.nodeIsPresent(node.body)
       ) {
-        if (
-          node.kind === SyntaxKind.MethodDeclaration &&
-          // @ts-expect-error -- internal?
-          !ts.nodeIsPresent(node.body)
-        ) {
-          throwErrorOnTsNode(
-            modifier,
-            "A decorator can only decorate a method implementation, not an overload."
-          );
-        } else {
-          throwErrorOnTsNode(modifier, "Decorators are not valid here.");
-        }
-      } else if (
-        legacyDecorators &&
-        (node.kind === SyntaxKind.GetAccessor ||
-          node.kind === SyntaxKind.SetAccessor)
-      ) {
-        // @ts-expect-error -- internal?
-        const accessors = ts.getAllAccessorDeclarations(
-          node.parent.members,
-          node
+        throwErrorOnTsNode(
+          modifier,
+          "A decorator can only decorate a method implementation, not an overload."
         );
-        if (
-          // @ts-expect-error -- internal?
-          ts.hasDecorators(accessors.firstAccessor) &&
-          node === accessors.secondAccessor
-        ) {
-          throwErrorOnTsNode(
-            modifier,
-            "Decorators cannot be applied to multiple get/set accessors of the same name."
-          );
-        }
       }
+
+      throwErrorOnTsNode(modifier, "Decorators are not valid here.");
     }
 
     if (modifier.kind !== SyntaxKind.ReadonlyKeyword) {