+++ src/plugins/html.js
@@ -1,0 +1,2 @@
+export * from "../language-html/index.js";
+export * as default from "../language-html/index.js";